requirements.txt
streamlit
psycopg2
mysql-connector-python
python-dotenv
pandas
pytz
